backend/requirements.txt
Django>=4.0
djangorestframework
markdown
django-filter
psycopg>=3.1.8